requirements.txt
pandas 
pandasai
streamlit
langchain_groq
python-dotenv